feat: Merge search command into list command with --query flag
This commit refactors the task management CLI by integrating the 'search'
command's functionality directly into the 'list' command via a new '--query'
flag. This change streamlines the user experience, reduces code duplication,
and simplifies the overall command structure.

Key changes include:

- **Core Logic Consolidation**: The core search logic has been moved from
  `internal/core/search.go` to `internal/core/list.go`. The `ListTasksParams`
  struct now includes a `Query` field, enabling the `FilterTasks` function
  to perform content-based searches alongside existing filters.
- **CLI Simplification**: The standalone `backlog search` command has been
  removed. Users can now perform searches using `backlog list --query <term>`.
  The `backlog list` command's example usage has been updated accordingly.
- **Interface Alignment**: The `TaskStore` interface in both `internal/cmd/root.go`
  and `internal/mcp/server.go` has been updated to remove the now-redundant
  `Search` method, ensuring consistency across the application's interfaces.
- **Test Refactoring**: All tests related to the `search` command have been
  merged into their respective `list` command test files (`internal/core/list_test.go`
  and `internal/cmd/task_list_test.go`). Redundant test files have been deleted.
  Several test-related issues, including undefined references and an incorrect
  `DependedOn` filter logic, were also addressed and fixed.
- **Documentation Updates**: `README.md` and `internal/mcp/prompt.md` have been
  revised to reflect the new `--query` flag for the `list` command, the removal
  of the `search` command, and updated usage examples for AI agents.

This change improves the usability and maintainability of the backlog CLI.
