+++ README.md
@@ -75,7 +75,7 @@ The rest of this document is mostly generated by AI (except the Inspiration sect
 
 - **Task Management**: Create, edit, list, and view tasks with rich metadata
 - **Hierarchical Structure**: Support for parent-child-grandchild task relationships (T01 → T01.01 → T01.01.01)
-- **Search & Filter**: Find tasks by content, status, parent relationships, and labels
+- **Search & Filter**: Find tasks by content, status, parent relationships, and labels using the `list` command with the `--query` flag
 - **AI-Friendly**: MCP server integration for seamless AI agent collaboration
 - **Git Integration**: Tasks are stored as Markdown files with automatic Git commits
 - **Offline-First**: Works completely offline with local Git repository storage
@@ -210,7 +210,6 @@ Backlog includes a Model Context Protocol (MCP) server that exposes task managem
 - `task_list`: List and filter tasks
 - `task_view`: Get detailed task information
 - `task_edit`: Update existing tasks
-- `task_search`: Search tasks by content
 - `task_archive`: Archive tasks so they are not displayed anymore but still lives in the repository.
 
 ### Usage with AI Agents
@@ -311,8 +310,8 @@ backlog list --limit 5 --offset 10              # Tasks 11-15
 backlog list --status "todo" --limit 3          # First 3 todo tasks
 
 # Search with pagination
-backlog search "api" --limit 5                  # First 5 API-related tasks
-backlog search "bug" --limit 3 --offset 5       # Search results 6-8
+backlog list --query "api" --limit 5                  # First 5 API-related tasks
+backlog list --query "bug" --limit 3 --offset 5       # Search results 6-8
 
 # View specific task
 backlog view T01.02
